refactor: split hypothesis evaluators
