+++ docs/agent-maintainer-hardening-branch-roadmap.md
@@ -1,6 +1,6 @@
 # Agent Maintainer Hardening Branch Roadmap
 
-Current chunk: `refactor/report-query-family`
+Current chunk: `refactor/report-hypothesis-evaluators`
 
 ## Goal
 
@@ -119,6 +119,8 @@ module boundary and keep behavior covered by focused tests.
     splitting the larger evaluator engine.
   - [x] Extract filtered query and recommendation reports so hypothesis
     evaluators can depend on a concrete module instead of the facade.
+  - [x] Extract hypothesis evaluators, result metrics, and shared agentic
+    evidence compaction into independently sized modules.
   - [ ] Extract investigation, hypothesis, walk/export, and query/recommendation
     families until the facade is below the configured thresholds.
 - Split `src/codex_usage_tracker/store/content_index.py` by ingestion, FTS,
